+++ shell/components/DrawerVisibilities.qml
@@ -9,7 +9,6 @@ PersistentProperties {
     property bool island
     property bool dashboard
     property bool settings
-    property bool wallhaven
     property bool obsidian
     property bool utilities
     property bool sidebar
@@ -21,7 +20,6 @@ PersistentProperties {
         launcher = false;
         island = false;
         dashboard = false;
-        wallhaven = false;
         obsidian = false;
         utilities = false;
         sidebar = false;